+++ surefilter-ui/prisma/migrations/20260302152137_add_default_seo_meta/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "SiteSettings" ADD COLUMN     "defaultMetaDesc" TEXT,
+ADD COLUMN     "defaultMetaKeywords" TEXT,
+ADD COLUMN     "defaultMetaTitle" TEXT,
+ADD COLUMN     "defaultMetaTitleSuffix" TEXT;